+++ ValidationNetwork/DnnHelpers.h
@@ -14,6 +14,6 @@
 #include "ActivationFunctions.h"
 
 std::tuple<std::vector<Matrix>, std::vector<std::vector<float>> > BackwardPropogation(Matrix in, Matrix y, std::vector<Matrix> w, std::vector<std::vector<float>> b,
-	std::vector<Matrix> A, std::vector<Matrix> Z, std::unordered_set<int> res, int learning_rate);
+	std::vector<Matrix> A, std::vector<Matrix> Z, std::unordered_set<int> res, float learning_rate);
 float DNN_Accuracy(Matrix last_activation, std::vector<float> labels);
 std::tuple<Matrix, Matrix, std::vector<float>> ShuffleInput(Matrix in, Matrix y, std::vector<float> labels);
